Crate rustc_infer

This crates defines the type inference engine.

Type inference. The type inference code can be found in the infer module; this code handles low-level equality and subtyping operations. The type check pass in the compiler is found in the rustc_typeck crate.

For more information about how rustc works, see the rustc dev guide.

Note

This API is completely unstable and subject to change.

Modules

errors 🔒

infer

traits

Trait Resolution. See the rustc-dev-guide for more information on how this works.