Struct rustc_infer::infer::at::At

pub struct At<'a, 'tcx> {
    pub infcx: &'a InferCtxt<'tcx>,
    pub cause: &'a ObligationCause<'tcx>,
    pub param_env: ParamEnv<'tcx>,
    pub define_opaque_types: bool,
}

Fields
infcx: &'a InferCtxt<'tcx>cause: &'a ObligationCause<'tcx>param_env: ParamEnv<'tcx>define_opaque_types: bool

Whether we should define opaque types or just treat them opaquely. Currently only used to prevent predicate matching from matching anything against opaque types.

Implementations

impl<'a, 'tcx> At<'a, 'tcx>

pub fn define_opaque_types(self, define_opaque_types: bool) -> Self

pub fn eq_impl_headers(
self,
expected: &ImplHeader<'tcx>,
actual: &ImplHeader<'tcx>
) -> InferResult<'tcx, ()>

Hacky routine for equating two impl headers in coherence.

pub fn sub_exp<T>(self, a_is_expected: bool, a: T, b: T) -> InferResult<'tcx, ()>where
T: ToTrace<'tcx>,

Makes a <: b, where a may or may not be expected.

See At::trace_exp and Trace::sub for a version of this method that only requires T: Relate<'tcx>

pub fn sup<T>(self, expected: T, actual: T) -> InferResult<'tcx, ()>where
T: ToTrace<'tcx>,

Makes actual <: expected. For example, if type-checking a call like foo(x), where foo: fn(i32), you might have sup(i32, x), since the “expected” type is the type that appears in the signature.

See At::trace and Trace::sub for a version of this method that only requires T: Relate<'tcx>

pub fn sub<T>(self, expected: T, actual: T) -> InferResult<'tcx, ()>where
T: ToTrace<'tcx>,

Makes expected <: actual.

See At::trace and Trace::sub for a version of this method that only requires T: Relate<'tcx>

pub fn eq_exp<T>(self, a_is_expected: bool, a: T, b: T) -> InferResult<'tcx, ()>where
T: ToTrace<'tcx>,

Makes expected <: actual.

See At::trace_exp and Trace::eq for a version of this method that only requires T: Relate<'tcx>

pub fn eq<T>(self, expected: T, actual: T) -> InferResult<'tcx, ()>where
T: ToTrace<'tcx>,

Makes expected <: actual.

See At::trace and Trace::eq for a version of this method that only requires T: Relate<'tcx>

pub fn relate<T>(
self,
expected: T,
variance: Variance,
actual: T
) -> InferResult<'tcx, ()>where
T: ToTrace<'tcx>,

pub fn lub<T>(self, expected: T, actual: T) -> InferResult<'tcx, T>where
T: ToTrace<'tcx>,

Computes the least-upper-bound, or mutual supertype, of two values. The order of the arguments doesn’t matter, but since this can result in an error (e.g., if asked to compute LUB of u32 and i32), it is meaningful to call one of them the “expected type”.

See At::trace and Trace::lub for a version of this method that only requires T: Relate<'tcx>

pub fn glb<T>(self, expected: T, actual: T) -> InferResult<'tcx, T>where
T: ToTrace<'tcx>,

Computes the greatest-lower-bound, or mutual subtype, of two values. As with lub order doesn’t matter, except for error cases.

See At::trace and Trace::glb for a version of this method that only requires T: Relate<'tcx>

pub fn trace<T>(self, expected: T, actual: T) -> Trace<'a, 'tcx>where
T: ToTrace<'tcx>,

Sets the “trace” values that will be used for error-reporting, but doesn’t actually perform any operation yet (this is useful when you want to set the trace using distinct values from those you wish to operate upon).

pub fn trace_exp<T>(self, a_is_expected: bool, a: T, b: T) -> Trace<'a, 'tcx>where
T: ToTrace<'tcx>,

Like trace, but the expected value is determined by the boolean argument (if true, then the first argument a is the “expected” value).