rustc_infer::infer::canonical

Module instantiate

This module contains code to instantiate new values into a Canonical<'tcx, T>.

For an overview of what canonicalization is and how it fits into rustc, check out the chapter in the rustc dev guide.

Structs

CanonicalInstantiator 🔒

Replaces the bound vars in a canonical binder with var values.

Traits

CanonicalExt

FIXME(-Znext-solver): This or public because it is shared with the new trait solver implementation. We should deduplicate canonicalization.

Functions

highest_var_in_clauses 🔒

instantiate_value 🔒

Instantiate the values from var_values into value. var_values must be values for the set of canonical variables that appear in value.