Enum rustc_infer::infer::canonical::ir::TyKind

pub enum TyKind<I>
where
    I: Interner,
{

    Bool,
    Char,
    Int(IntTy),
    Uint(UintTy),
    Float(FloatTy),
    Adt(<I as Interner>::AdtDef, <I as Interner>::GenericArgs),
    Foreign(<I as Interner>::DefId),
    Str,
    Array(<I as Interner>::Ty, <I as Interner>::Const),
    Pat(<I as Interner>::Ty, <I as Interner>::Pat),
    Slice(<I as Interner>::Ty),
    RawPtr(<I as Interner>::Ty, Mutability),
    Ref(<I as Interner>::Region, <I as Interner>::Ty, Mutability),
    FnDef(<I as Interner>::DefId, <I as Interner>::GenericArgs),
    FnPtr(Binder<I, FnSig<I>>),
    Dynamic(<I as Interner>::BoundExistentialPredicates, <I as Interner>::Region, DynKind),
    Closure(<I as Interner>::DefId, <I as Interner>::GenericArgs),
    CoroutineClosure(<I as Interner>::DefId, <I as Interner>::GenericArgs),
    Coroutine(<I as Interner>::DefId, <I as Interner>::GenericArgs),
    CoroutineWitness(<I as Interner>::DefId, <I as Interner>::GenericArgs),
    Never,
    Tuple(<I as Interner>::Tys),
    Alias(AliasTyKind, AliasTy<I>),
    Param(<I as Interner>::ParamTy),
    Bound(DebruijnIndex, <I as Interner>::BoundTy),
    Placeholder(<I as Interner>::PlaceholderTy),
    Infer(InferTy),
    Error(<I as Interner>::ErrorGuaranteed),

}

Defines the kinds of types used by the type system.

Types written by the user start out as hir::TyKind and get converted to this representation using <dyn HirTyLowerer>::lower_ty.

Variants

Bool

The primitive boolean type. Written as bool.

Char

The primitive character type; holds a Unicode scalar value (a non-surrogate code point). Written as char.

Int(IntTy)

A primitive signed integer type. For example, i32.

Uint(UintTy)

A primitive unsigned integer type. For example, u32.

Float(FloatTy)

A primitive floating-point type. For example, f64.

Adt(<I as Interner>::AdtDef, <I as Interner>::GenericArgs)

Algebraic data types (ADT). For example: structures, enumerations and unions.

For example, the type List<i32> would be represented using the AdtDef for struct List<T> and the args [i32].

Note that generic parameters in fields only get lazily instantiated by using something like adt_def.all_fields().map(|field| field.ty(interner, args)).

Foreign(<I as Interner>::DefId)

An unsized FFI type that is opaque to Rust. Written as extern type T.

Str

The pointee of a string slice. Written as str.

Array(<I as Interner>::Ty, <I as Interner>::Const)

An array with the given length. Written as [T; N].

Pat(<I as Interner>::Ty, <I as Interner>::Pat)

A pattern newtype. Takes any type and restricts its valid values to its pattern. This will also change the layout to take advantage of this restriction. Only Copy and Clone will automatically get implemented for pattern types. Auto-traits treat this as if it were an aggregate with a single nested type. Only supports integer range patterns for now.

Slice(<I as Interner>::Ty)

The pointee of an array slice. Written as [T].

RawPtr(<I as Interner>::Ty, Mutability)

A raw pointer. Written as *mut T or *const T

Ref(<I as Interner>::Region, <I as Interner>::Ty, Mutability)

A reference; a pointer with an associated lifetime. Written as &'a mut T or &'a T.

FnDef(<I as Interner>::DefId, <I as Interner>::GenericArgs)

The anonymous type of a function declaration/definition. Each function has a unique type.

For the function fn foo() -> i32 { 3 } this type would be shown to the user as fn() -> i32 {foo}.

For example the type of bar here:

fn foo() -> i32 { 1 }
let bar = foo; // bar: fn() -> i32 {foo}

FnPtr(Binder<I, FnSig<I>>)

A pointer to a function. Written as fn() -> i32.

Note that both functions and closures start out as either FnDef or Closure which can be then be coerced to this variant.

For example the type of bar here:

fn foo() -> i32 { 1 }
let bar: fn() -> i32 = foo;

Dynamic(<I as Interner>::BoundExistentialPredicates, <I as Interner>::Region, DynKind)

A trait object. Written as dyn for<'b> Trait<'b, Assoc = u32> + Send + 'a.

Closure(<I as Interner>::DefId, <I as Interner>::GenericArgs)

The anonymous type of a closure. Used to represent the type of |a| a.

Closure args contain both the - potentially instantiated - generic parameters of its parent and some synthetic parameters. See the documentation for ClosureArgs for more details.

CoroutineClosure(<I as Interner>::DefId, <I as Interner>::GenericArgs)

The anonymous type of a closure. Used to represent the type of async |a| a.

Coroutine-closure args contain both the - potentially instantiated - generic parameters of its parent and some synthetic parameters. See the documentation for CoroutineClosureArgs for more details.

Coroutine(<I as Interner>::DefId, <I as Interner>::GenericArgs)

The anonymous type of a coroutine. Used to represent the type of |a| yield a.

For more info about coroutine args, visit the documentation for CoroutineArgs.

CoroutineWitness(<I as Interner>::DefId, <I as Interner>::GenericArgs)

A type representing the types stored inside a coroutine. This should only appear as part of the CoroutineArgs.

Unlike upvars, the witness can reference lifetimes from inside of the coroutine itself. To deal with them in the type of the coroutine, we convert them to higher ranked lifetimes bound by the witness itself.

This contains the DefId and the GenericArgsRef of the coroutine. The actual witness types are computed on MIR by the mir_coroutine_witnesses query.

Looking at the following example, the witness for this coroutine may end up as something like for<'a> [Vec<i32>, &'a Vec<i32>]:

#![feature(coroutines)]
#[coroutine] static |a| {
    let x = &vec![3];
    yield a;
    yield x[0];
}

Never

The never type !.

Tuple(<I as Interner>::Tys)

A tuple type. For example, (i32, bool).

Alias(AliasTyKind, AliasTy<I>)

A projection, opaque type, weak type alias, or inherent associated type. All of these types are represented as pairs of def-id and args, and can be normalized, so they are grouped conceptually.

Param(<I as Interner>::ParamTy)

A type parameter; for example, T in fn f<T>(x: T) {}.

Bound(DebruijnIndex, <I as Interner>::BoundTy)

Bound type variable, used to represent the 'a in for<'a> fn(&'a ()).

For canonical queries, we replace inference variables with bound variables, so e.g. when checking whether &'_ (): Trait<_> holds, we canonicalize that to for<'a, T> &'a (): Trait<T> and then convert the introduced bound variables back to inference variables in a new inference context when inside of the query.

It is conventional to render anonymous bound types like ^N or ^D_N, where N is the bound variable’s anonymous index into the binder, and D is the debruijn index, or totally omitted if the debruijn index is zero.

See the rustc-dev-guide for more details about higher-ranked trait bounds and canonical queries.

Placeholder(<I as Interner>::PlaceholderTy)

A placeholder type, used during higher ranked subtyping to instantiate bound variables.

It is conventional to render anonymous placeholder types like !N or !U_N, where N is the placeholder variable’s anonymous index (which corresponds to the bound variable’s index from the binder from which it was instantiated), and U is the universe index in which it is instantiated, or totally omitted if the universe index is zero.

Infer(InferTy)

A type variable used during type checking.

Similar to placeholders, inference variables also live in a universe to correctly deal with higher ranked types. Though unlike placeholders, that universe is stored in the InferCtxt instead of directly inside of the type.

Error(<I as Interner>::ErrorGuaranteed)

A placeholder for a type which could not be computed; this is propagated to avoid useless error messages.

Implementations

impl<I> TyKind<I>
where I: Interner,

pub fn is_primitive(&self) -> bool