rustc_infer::infer::canonical::ir
Module solve

Modules

inspect

Data structure used to inspect trait solver behavior.

Structs

ExternalConstraintsData

Additional constraints returned on success.

Goal

A goal is a statement, i.e. predicate, we want to prove given some assumptions, i.e. param_env.

NestedNormalizationGoals

NoSolution

PredefinedOpaquesData

Opaques that are defined in the inference context before a query is called.

QueryInput

Response

Enums

AdtDestructorKind

Indicates that a impl Drop for Adt is const or not.

BuiltinImplSource

CandidateSource

Possible ways the given goal can be proven.

Certainty

GoalSource

Why a specific goal has to be proven.

MaybeCause

Why we failed to evaluate a goal.

Type Aliases

CanonicalInput

CanonicalResponse

QueryResult

The result of evaluating a canonical query.