rustc_infer::infer::canonical::ir
Struct Binder

pub struct Binder<I, T>
where
    I: Interner,
{
    value: T,
    bound_vars: <I as Interner>::BoundVarKinds,
}

Binder is a binder for higher-ranked lifetimes or types. It is part of the compiler’s representation for things like for<'a> Fn(&'a isize) (which would be represented by the type PolyTraitRef == Binder<I, TraitRef>). Note that when we instantiate, erase, or otherwise “discharge” these bound vars, we change the type from Binder<I, T> to just T (see e.g., liberate_late_bound_regions).

Decodable and Encodable are implemented for Binder<T> using the impl_binder_encode_decode! macro.

Fields
value: Tbound_vars: <I as Interner>::BoundVarKinds
Implementations

impl<I, T> Binder<I, T>
where I: Interner, T: TypeVisitable<I>,

pub fn dummy(value: T) -> Binder<I, T>

Wraps value in a binder, asserting that value does not contain any bound vars that would be bound by the binder. This is commonly used to ‘inject’ a value T into a different binding level.

pub fn bind_with_vars( value: T, bound_vars: <I as Interner>::BoundVarKinds, ) -> Binder<I, T>

impl<I, T> Binder<I, T>
where I: Interner,

pub fn skip_binder(self) -> T

Skips the binder and returns the “bound” value. This is a risky thing to do because it’s easy to get confused about De Bruijn indices and the like. It is usually better to discharge the binder using no_bound_vars or instantiate_bound_regions or something like that. skip_binder is only valid when you are either extracting data that has nothing to do with bound vars, you are doing some sort of test that does not involve bound regions, or you are being very careful about your depth accounting.

Some examples where skip_binder is reasonable:

extracting the DefId from a PolyTraitRef;

comparing the self type of a PolyTraitRef to see if it is equal to a type parameter X, since the type X does not reference any regions

pub fn bound_vars(&self) -> <I as Interner>::BoundVarKinds

pub fn as_ref(&self) -> Binder<I, &T>

pub fn as_deref(&self) -> Binder<I, &<T as Deref>::Target>
where T: Deref,

pub fn map_bound_ref<F, U>(&self, f: F) -> Binder<I, U>
where U: TypeVisitable<I>, F: FnOnce(&T) -> U,

pub fn map_bound<F, U>(self, f: F) -> Binder<I, U>
where U: TypeVisitable<I>, F: FnOnce(T) -> U,

pub fn try_map_bound<F, U, E>(self, f: F) -> Result<Binder<I, U>, E>
where U: TypeVisitable<I>, F: FnOnce(T) -> Result<U, E>,

pub fn rebind<U>(&self, value: U) -> Binder<I, U>
where U: TypeVisitable<I>,

Wraps a value in a binder, using the same bound variables as the current Binder. This should not be used if the new value changes the bound variables. Note: the (old or new) value itself does not necessarily need to name all the bound variables.

This currently doesn’t do anything different than bind, because we don’t actually track bound vars. However, semantically, it is different because bound vars aren’t allowed to change here, whereas they are in bind. This may be (debug) asserted in the future.

pub fn no_bound_vars(self) -> Option<T>
where T: TypeVisitable<I>,

Unwraps and returns the value within, but only if it contains no bound vars at all. (In other words, if this binder – and indeed any enclosing binder – doesn’t bind anything at all.) Otherwise, returns None.

(One could imagine having a method that just unwraps a single binder, but permits late-bound vars bound by enclosing binders, but that would require adjusting the debruijn indices, and given the shallow binding structure we often use, would not be that useful.)

impl<I, T> Binder<I, Option<T>>
where I: Interner,

pub fn transpose(self) -> Option<Binder<I, T>>

impl<I, T> Binder<I, T>
where I: Interner, T: IntoIterator,

pub fn iter(self) -> impl Iterator<Item = Binder<I, <T as IntoIterator>::Item>>

impl<I> Binder<I, TraitRef<I>>
where I: Interner,

pub fn self_ty(&self) -> Binder<I, <I as Interner>::Ty>

pub fn def_id(&self) -> <I as Interner>::DefId

impl<I> Binder<I, TraitPredicate<I>>
where I: Interner,

pub fn def_id(self) -> <I as Interner>::DefId

pub fn self_ty(self) -> Binder<I, <I as Interner>::Ty>

pub fn polarity(self) -> PredicatePolarity

impl<I> Binder<I, ExistentialPredicate<I>>
where I: Interner,

pub fn with_self_ty( &self, cx: I, self_ty: <I as Interner>::Ty, ) -> <I as Interner>::Clause

Given an existential predicate like ?Self: PartialEq<u32> (e.g., derived from dyn PartialEq<u32>), and a concrete type self_ty, returns a full predicate where the existentially quantified variable ?Self has been replaced with self_ty (e.g., self_ty: PartialEq<u32>, in our example).

impl<I> Binder<I, ExistentialTraitRef<I>>
where I: Interner,

pub fn def_id(&self) -> <I as Interner>::DefId

pub fn with_self_ty( &self, cx: I, self_ty: <I as Interner>::Ty, ) -> Binder<I, TraitRef<I>>

Object types don’t have a self type specified. Therefore, when we convert the principal trait-ref into a normal trait-ref, you must give some self type. A common choice is mk_err() or some placeholder type.

impl<I> Binder<I, ExistentialProjection<I>>
where I: Interner,

pub fn with_self_ty( &self, cx: I, self_ty: <I as Interner>::Ty, ) -> Binder<I, ProjectionPredicate<I>>

pub fn item_def_id(&self) -> <I as Interner>::DefId

impl<I> Binder<I, ProjectionPredicate<I>>
where I: Interner,

pub fn trait_def_id(&self, cx: I) -> <I as Interner>::DefId

Returns the DefId of the trait of the associated item being projected.

pub fn required_poly_trait_ref(&self, cx: I) -> Binder<I, TraitRef<I>>

Get the trait ref required for this projection to be well formed. Note that for generic associated types the predicates of the associated type also need to be checked.

pub fn term(&self) -> Binder<I, <I as Interner>::Term>

pub fn projection_def_id(&self) -> <I as Interner>::DefId

The DefId of the TraitItem for the associated type.

Note that this is not the DefId of the TraitRef containing this associated type, which is in tcx.associated_item(projection_def_id()).container.

impl<I> Binder<I, FnSig<I>>
where I: Interner,

pub fn inputs(self) -> Binder<I, <I as Interner>::FnInputTys>

pub fn input(self, index: usize) -> Binder<I, <I as Interner>::Ty>

pub fn inputs_and_output(self) -> Binder<I, <I as Interner>::Tys>

pub fn output(self) -> Binder<I, <I as Interner>::Ty>

pub fn c_variadic(self) -> bool

pub fn safety(self) -> <I as Interner>::Safety

pub fn abi(self) -> <I as Interner>::Abi

pub fn is_fn_trait_compatible(&self) -> bool

pub fn split(self) -> (Binder<I, FnSigTys<I>>, FnHeader<I>)

impl<I> Binder<I, FnSigTys<I>>
where I: Interner,

pub fn with(self, hdr: FnHeader<I>) -> Binder<I, FnSig<I>>

pub fn inputs(self) -> Binder<I, <I as Interner>::FnInputTys>

pub fn input(self, index: usize) -> Binder<I, <I as Interner>::Ty>

pub fn inputs_and_output(self) -> Binder<I, <I as Interner>::Tys>

pub fn output(self) -> Binder<I, <I as Interner>::Ty>