Struct rustc_infer::infer::canonical::ir::TypeFlags

pub struct TypeFlags(InternalBitFlags);

Flags that we track on types. These flags are propagated upwards through the type during type construction, so that we can quickly check whether the type has various kinds of types in it without recursing over the type itself.

Tuple Fields
0: InternalBitFlags
Implementations

impl TypeFlags

pub const HAS_TY_PARAM: TypeFlags = _

Does this have Param?

pub const HAS_RE_PARAM: TypeFlags = _

Does this have ReEarlyParam?

pub const HAS_CT_PARAM: TypeFlags = _

Does this have ConstKind::Param?

pub const HAS_PARAM: TypeFlags = _

pub const HAS_TY_INFER: TypeFlags = _

Does this have Infer?

pub const HAS_RE_INFER: TypeFlags = _

Does this have ReVar?

pub const HAS_CT_INFER: TypeFlags = _

Does this have ConstKind::Infer?

pub const HAS_INFER: TypeFlags = _

Does this have inference variables? Used to determine whether inference is required.

pub const HAS_TY_PLACEHOLDER: TypeFlags = _

Does this have Placeholder?

pub const HAS_RE_PLACEHOLDER: TypeFlags = _

Does this have RePlaceholder?

pub const HAS_CT_PLACEHOLDER: TypeFlags = _

Does this have ConstKind::Placeholder?

pub const HAS_PLACEHOLDER: TypeFlags = _

Does this have placeholders?

pub const HAS_FREE_LOCAL_REGIONS: TypeFlags = _

true if there are “names” of regions and so forth that are local to a particular fn/inferctxt

pub const HAS_FREE_LOCAL_NAMES: TypeFlags = _

true if there are “names” of types and regions and so forth that are local to a particular fn

pub const HAS_TY_PROJECTION: TypeFlags = _

Does this have Projection?

pub const HAS_TY_WEAK: TypeFlags = _

Does this have Weak?

pub const HAS_TY_OPAQUE: TypeFlags = _

Does this have Opaque?

pub const HAS_TY_INHERENT: TypeFlags = _

Does this have Inherent?

pub const HAS_CT_PROJECTION: TypeFlags = _

Does this have ConstKind::Unevaluated?

pub const HAS_ALIAS: TypeFlags = _

Does this have Alias or ConstKind::Unevaluated?

Rephrased, could this term be normalized further?

pub const HAS_ERROR: TypeFlags = _

Is an error type/lifetime/const reachable?

pub const HAS_FREE_REGIONS: TypeFlags = _

Does this have any region that “appears free” in the type? Basically anything but ReBound and ReErased.

pub const HAS_RE_BOUND: TypeFlags = _

Does this have any ReBound regions?

pub const HAS_TY_BOUND: TypeFlags = _

Does this have any Bound types?

pub const HAS_CT_BOUND: TypeFlags = _

Does this have any ConstKind::Bound consts?

pub const HAS_BOUND_VARS: TypeFlags = _

Does this have any bound variables? Used to check if a global bound is safe to evaluate.

pub const HAS_RE_ERASED: TypeFlags = _

Does this have any ReErased regions?

pub const STILL_FURTHER_SPECIALIZABLE: TypeFlags = _

Does this value have parameters/placeholders/inference variables which could be replaced later, in a way that would change the results of impl specialization?

pub const HAS_TY_FRESH: TypeFlags = _

Does this value have InferTy::FreshTy/FreshIntTy/FreshFloatTy?

pub const HAS_CT_FRESH: TypeFlags = _

Does this value have InferConst::Fresh?

pub const HAS_TY_COROUTINE: TypeFlags = _

Does this have Coroutine or CoroutineWitness?

pub const HAS_BINDER_VARS: TypeFlags = _

Does this have any binders with bound vars (e.g. that need to be anonymized)?

impl TypeFlags

pub const fn empty() -> TypeFlags

Get a flags value with all bits unset.

pub const fn all() -> TypeFlags

Get a flags value with all known bits set.

pub const fn bits(&self) -> u32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u32) -> Option<TypeFlags>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> TypeFlags

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> TypeFlags

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<TypeFlags>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: TypeFlags) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: TypeFlags) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: TypeFlags)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: TypeFlags)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: TypeFlags)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: TypeFlags, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: TypeFlags) -> TypeFlags

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: TypeFlags) -> TypeFlags

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: TypeFlags) -> TypeFlags

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: TypeFlags) -> TypeFlags

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> TypeFlags

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl TypeFlags

pub const fn iter(&self) -> Iter<TypeFlags>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<TypeFlags>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.