Module rustc_infer::infer::canonical::query_response

This module contains the code to instantiate a “query result”, and in particular to extract out the resulting region obligations and encode them therein.

For an overview of what canonicalization is and how it fits into rustc, check out the chapter in the rustc dev guide.

Functions

make_query_region_constraints

Given the region obligations and constraints scraped from the infcx, creates query region constraints.