Enum rustc_infer::infer::error_reporting::need_type_info::InferSourceKind

enum InferSourceKind<'tcx> {
    LetBinding {
        insert_span: Span,
        pattern_name: Option<Ident>,
        ty: Ty<'tcx>,
    },
    ClosureArg {
        insert_span: Span,
        ty: Ty<'tcx>,
    },
    GenericArg {
        insert_span: Span,
        argument_index: usize,
        generics_def_id: DefId,
        def_id: DefId,
        generic_args: &'tcx [GenericArg<'tcx>],
    },
    FullyQualifiedMethodCall {
        receiver: &'tcx Expr<'tcx>,
        successor: (&'static str, BytePos),
        substs: SubstsRef<'tcx>,
        def_id: DefId,
    },
    ClosureReturn {
        ty: Ty<'tcx>,
        data: &'tcx FnRetTy<'tcx>,
        should_wrap_expr: Option<Span>,
    },
}

Variants

LetBinding

Fields

insert_span: Span

pattern_name: Option<Ident>

ty: Ty<'tcx>

ClosureArg

Fields

insert_span: Span

ty: Ty<'tcx>

GenericArg

Fields

insert_span: Span

argument_index: usize

generics_def_id: DefId

def_id: DefId

generic_args: &'tcx [GenericArg<'tcx>]

FullyQualifiedMethodCall

Fields

receiver: &'tcx Expr<'tcx>

successor: (&'static str, BytePos)

If the method has other arguments, this is “, “ and the start of the first argument, while for methods without arguments this is “)” and the end of the method call.

substs: SubstsRef<'tcx>

def_id: DefId

ClosureReturn

Fields

ty: Ty<'tcx>

data: &'tcx FnRetTy<'tcx>

should_wrap_expr: Option<Span>

Implementations

impl<'tcx> InferSourceKind<'tcx>

fn ty_localized_msg(&self, infcx: &InferCtxt<'_, 'tcx>) -> (&'static str, String)