Struct rustc_infer::infer::error_reporting::need_type_info::FindInferSourceVisitor

struct FindInferSourceVisitor<'a, 'tcx> {
    infcx: &'a InferCtxt<'tcx>,
    typeck_results: &'a TypeckResults<'tcx>,
    target: GenericArg<'tcx>,
    attempt: usize,
    infer_source_cost: usize,
    infer_source: Option<InferSource<'tcx>>,
}

A visitor which searches for the “best” spot to use in the inference error.

For this it walks over the hir body and tries to check all places where inference variables could be bound.

While doing so, the currently best spot is stored in infer_source. For details on how we rank spots, see Self::source_cost

Fields
infcx: &'a InferCtxt<'tcx>typeck_results: &'a TypeckResults<'tcx>target: GenericArg<'tcx>attempt: usizeinfer_source_cost: usizeinfer_source: Option<InferSource<'tcx>>
Implementations

impl<'a, 'tcx> FindInferSourceVisitor<'a, 'tcx>

fn new(
infcx: &'a InferCtxt<'tcx>,
typeck_results: &'a TypeckResults<'tcx>,
target: GenericArg<'tcx>
) -> Self

fn source_cost(&self, source: &InferSource<'tcx>) -> usize

Computes cost for the given source.

Sources with a small cost are prefer and should result in a clearer and idiomatic suggestion.

fn update_infer_source(&mut self, new_source: InferSource<'tcx>)

Uses fn source_cost to determine whether this inference source is preferable to previous sources. We generally prefer earlier sources.

fn node_substs_opt(&self, hir_id: HirId) -> Option<SubstsRef<'tcx>>

fn opt_node_type(&self, hir_id: HirId) -> Option<Ty<'tcx>>

fn generic_arg_is_target(&self, arg: GenericArg<'tcx>) -> bool

fn generic_arg_contains_target(&self, arg: GenericArg<'tcx>) -> bool

Does this generic argument contain our target inference variable in a way which can be written by the user.

fn expr_inferred_subst_iter(
&self,
expr: &'tcx Expr<'tcx>
) -> Box<dyn Iterator<Item = InsertableGenericArgs<'tcx>> + 'a>

fn resolved_path_inferred_subst_iter(
&self,
path: &'tcx Path<'tcx>,
substs: SubstsRef<'tcx>
) -> impl Iterator<Item = InsertableGenericArgs<'tcx>> + 'a

fn path_inferred_subst_iter(
&self,
hir_id: HirId,
substs: SubstsRef<'tcx>,
qpath: &'tcx QPath<'tcx>
) -> Box<dyn Iterator<Item = InsertableGenericArgs<'tcx>> + 'a>