Function rustc_infer::infer::lexical_region_resolve::resolve

pub(crate) fn resolve<'tcx>(
    param_env: ParamEnv<'tcx>,
    region_rels: &RegionRelations<'_, 'tcx>,
    var_infos: VarInfos,
    data: RegionConstraintData<'tcx>
) -> (LexicalRegionResolutions<'tcx>, Vec<RegionResolutionError<'tcx>>)

This function performs lexical region resolution given a complete set of constraints and variable origins. It performs a fixed-point iteration to find region values which satisfy all constraints, assuming such values can be found. It returns the final values of all the variables as well as a set of errors that must be reported.