Module rustc_infer::infer::opaque_types

Modules

table 🔒

Structs

ConstrainOpaqueTypeRegionVisitor 🔒

OpaqueTypeDecl

Information about the opaque types whose values we are inferring in this function (these are the impl Trait that appear in the return type).

OpaqueTypeStorage

OpaqueTypeTable

Enums

UseKind

Functions

may_define_opaque_type 🔒

Returns true if opaque_hir_id is a sibling or a child of a sibling of def_id.

Type Definitions

OpaqueTypeMap