rustc_infer::infer::relate::type_relating
Struct TypeRelating

pub(crate) struct TypeRelating<'infcx, 'tcx> {
    infcx: &'infcx InferCtxt<'tcx>,
    trace: TypeTrace<'tcx>,
    param_env: ParamEnv<'tcx>,
    define_opaque_types: DefineOpaqueTypes,
    ambient_variance: Variance,
    obligations: Vec<PredicateObligation<'tcx>>,
    cache: DelayedSet<(Variance, Ty<'tcx>, Ty<'tcx>)>,
}

Enforce that a is equal to or a subtype of b.

Fields
infcx: &'infcx InferCtxt<'tcx>trace: TypeTrace<'tcx>param_env: ParamEnv<'tcx>define_opaque_types: DefineOpaqueTypesambient_variance: Varianceobligations: Vec<PredicateObligation<'tcx>>cache: DelayedSet<(Variance, Ty<'tcx>, Ty<'tcx>)>

The cache only tracks the ambient_variance as it’s the only field which is mutable and which meaningfully changes the result when relating types.

The cache does not track whether the state of the InferCtxt has been changed or whether we’ve added any obligations to self.goals. Whether a goal is added once or multiple times is not really meaningful.

Changes in the inference state may delay some type inference to the next fulfillment loop. Given that this loop is already necessary, this is also not a meaningful change. Consider the following three relations:

Vec<?0> sub Vec<?1>
?0 eq u32
Vec<?0> sub Vec<?1>

Without a cache, the second Vec<?0> sub Vec<?1> would eagerly constrain ?1 to u32. When using the cache entry from the first time we’ve related these types, this only happens when later proving the Subtype(?0, ?1) goal from the first relation.

Implementations

impl<'infcx, 'tcx> TypeRelating<'infcx, 'tcx>

pub(crate) fn new( infcx: &'infcx InferCtxt<'tcx>, trace: TypeTrace<'tcx>, param_env: ParamEnv<'tcx>, define_opaque_types: DefineOpaqueTypes, ambient_variance: Variance, ) -> TypeRelating<'infcx, 'tcx>

pub(crate) fn into_obligations(self) -> Vec<PredicateObligation<'tcx>>