Struct rustc_infer::infer::InferCtxtBuilder

pub struct InferCtxtBuilder<'tcx> {
    tcx: TyCtxt<'tcx>,
    defining_use_anchor: DefiningAnchor,
    considering_regions: bool,
    fresh_typeck_results: Option<RefCell<TypeckResults<'tcx>>>,
    normalize_fn_sig_for_diagnostic: Option<Lrc<dyn Fn(&InferCtxt<'_, 'tcx>, PolyFnSig<'tcx>) -> PolyFnSig<'tcx>>>,
}

A temporary returned by tcx.infer_ctxt(). This is necessary for multiple InferCtxt to share the same in_progress_typeck_results without using Rc or something similar.

Fields
tcx: TyCtxt<'tcx>defining_use_anchor: DefiningAnchorconsidering_regions: boolfresh_typeck_results: Option<RefCell<TypeckResults<'tcx>>>normalize_fn_sig_for_diagnostic: Option<Lrc<dyn Fn(&InferCtxt<'_, 'tcx>, PolyFnSig<'tcx>) -> PolyFnSig<'tcx>>>
Implementations

impl<'tcx> InferCtxtBuilder<'tcx>

pub fn with_fresh_in_progress_typeck_results(
self,
table_owner: LocalDefId
) -> Self

Used only by rustc_typeck during body type-checking/inference, will initialize in_progress_typeck_results with fresh TypeckResults. Will also change the scope for opaque type defining use checks to the given owner.

pub fn with_opaque_type_inference(
self,
defining_use_anchor: DefiningAnchor
) -> Self

Whenever the InferCtxt should be able to handle defining uses of opaque types, you need to call this function. Otherwise the opaque type will be treated opaquely.

It is only meant to be called in two places, for typeck (via with_fresh_in_progress_typeck_results) and for the inference context used in mir borrowck.

pub fn ignoring_regions(self) -> Self

pub fn with_normalize_fn_sig_for_diagnostic(
self,
fun: Lrc<dyn Fn(&InferCtxt<'_, 'tcx>, PolyFnSig<'tcx>) -> PolyFnSig<'tcx>>
) -> Self

pub fn enter_with_canonical<T, R>(
&mut self,
span: Span,
canonical: &Canonical<'tcx, T>,
f: impl for<'a> FnOnce(InferCtxt<'a, 'tcx>, T, CanonicalVarValues<'tcx>) -> R
) -> Rwhere
T: TypeFoldable<'tcx>,

Given a canonical value C as a starting point, create an inference context that contains each of the bound values within instantiated as a fresh variable. The f closure is invoked with the new infcx, along with the instantiated value V and a substitution S. This substitution S maps from the bound values in C to their instantiated values in V (in other words, S(C) = V).

pub fn enter<R>(&mut self, f: impl for<'a> FnOnce(InferCtxt<'a, 'tcx>) -> R) -> R