Trait rustc_infer::infer::ObligationEmittingRelation

pub trait ObligationEmittingRelation<'tcx>: TypeRelation<'tcx> {
    // Required methods
    fn register_obligations(&mut self, obligations: PredicateObligations<'tcx>);
    fn register_predicates(
        &mut self,
        obligations: impl IntoIterator<Item: ToPredicate<'tcx>>
    );
    fn alias_relate_direction(&self) -> AliasRelationDirection;

    // Provided method
    fn register_type_relate_obligation(&mut self, a: Ty<'tcx>, b: Ty<'tcx>) { ... }
}

Required Methods

fn register_obligations(&mut self, obligations: PredicateObligations<'tcx>)

Register obligations that must hold in order for this relation to hold

fn register_predicates( &mut self, obligations: impl IntoIterator<Item: ToPredicate<'tcx>> )

Register predicates that must hold in order for this relation to hold. Uses a default obligation cause, ObligationEmittingRelation::register_obligations should be used if control over the obligation causes is required.

fn alias_relate_direction(&self) -> AliasRelationDirection

Relation direction emitted for AliasRelate predicates, corresponding to the direction of the relation.

Provided Methods

fn register_type_relate_obligation(&mut self, a: Ty<'tcx>, b: Ty<'tcx>)

Register an obligation that both types must be related to each other according to the ty::AliasRelationDirection given by ObligationEmittingRelation::alias_relate_direction