rustc_infer::traits
Module solve

Modules

inspect

Data structure used to inspect trait solver behavior.

Structs

ExternalConstraints

ExternalConstraintsData

Additional constraints returned on success.

NestedNormalizationGoals

NoSolution

PredefinedOpaques

PredefinedOpaquesData

Opaques that are defined in the inference context before a query is called.

Response

Enums

BuiltinImplSource

Certainty

GoalSource

Why a specific goal has to be proven.

MaybeCause

Why we failed to evaluate a goal.

Reveal

Depending on the stage of compilation, we want projection to be more or less conservative.

SolverMode

Type Aliases

CandidateSource

CanonicalInput

CanonicalResponse

Goal

QueryInput

QueryResult