Type Alias rustc_infer::traits::solve::QueryResult

pub type QueryResult<'tcx> = Result<Canonical<TyCtxt<'tcx>, Response<'tcx>>, NoSolution>;

The result of evaluating a canonical query.

FIXME: We use a different type than the existing canonical queries. This is because we need to add a Certainty for overflow and may want to restructure this code without having to worry about changes to currently used code. Once we’ve made progress on this solver, merge the two responses again.

Aliased Type

enum QueryResult<'tcx> {
    Ok(Canonical<TyCtxt<'tcx>, Response<'tcx>>),
    Err(NoSolution),
}

Variants

Ok(Canonical<TyCtxt<'tcx>, Response<'tcx>>)

Contains the success value

Err(NoSolution)

Contains the error value

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 40 bytes

Size for each variant:

Ok: 40 bytes

Err: 0 bytes