rustc_infer::traits::specialization_graph
Struct Graph

pub struct Graph {
    pub parent: UnordMap<DefId, DefId>,
    pub children: UnordMap<DefId, Children>,
}

A per-trait graph of impls in specialization order. At the moment, this graph forms a tree rooted with the trait itself, with all other nodes representing impls, and parent-child relationships representing specializations.

The graph provides two key services:

Construction. This implicitly checks for overlapping impls (i.e., impls that overlap but where neither specializes the other – an artifact of the simple “chain” rule.

Parent extraction. In particular, the graph can give you the immediate parents of a given specializing impl, which is needed for extracting default items amongst other things. In the simple “chain” rule, every impl has at most one parent.

Fields
parent: UnordMap<DefId, DefId>

All impls have a parent; the “root” impls have as their parent the def_id of the trait.

children: UnordMap<DefId, Children>

The “root” impls are found by looking up the trait’s def_id.

Implementations

impl Graph

pub fn new() -> Graph

pub fn parent(&self, child: DefId) -> DefId

The parent of a given impl, which is the DefId of the trait when the impl is a “specialization root”.