Struct rustc_infer::traits::specialization_graph::LeafDef

pub struct LeafDef {
    pub item: AssocItem,
    pub defining_node: Node,
    pub finalizing_node: Option<Node>,
}

Information about the most specialized definition of an associated item.

Fields
item: AssocItem

The associated item described by this LeafDef.

defining_node: Node

The node in the specialization graph containing the definition of item.

finalizing_node: Option<Node>

The “top-most” (ie. least specialized) specialization graph node that finalized the definition of item.

Example:

#![feature(specialization)]
trait Tr {
    fn assoc(&self);
}

impl<T> Tr for T {
    default fn assoc(&self) {}
}

impl Tr for u8 {}

If we start the leaf definition search at impl Tr for u8, that impl will be the finalizing_node, while defining_node will be the generic impl.

If the leaf definition search is started at the generic impl, finalizing_node will be None, since the most specialized impl we found still allows overriding the method (doesn’t finalize it).