Type Alias rustc_infer::traits::TraitObligation

pub type TraitObligation<'tcx> = Obligation<'tcx, TraitPredicate<'tcx>>;

Aliased Type

struct TraitObligation<'tcx> {
    pub cause: ObligationCause<'tcx>,
    pub param_env: ParamEnv<'tcx>,
    pub predicate: TraitPredicate<'tcx>,
    pub recursion_depth: usize,
}

Fields
cause: ObligationCause<'tcx>

The reason we have to prove this thing.

param_env: ParamEnv<'tcx>

The environment in which we should prove this thing.

predicate: TraitPredicate<'tcx>

The thing we are trying to prove.

recursion_depth: usize

If we started proving this as a result of trying to prove something else, track the total depth to ensure termination. If this goes over a certain threshold, we abort compilation – in such cases, we can not say whether or not the predicate holds for certain. Stupid halting problem; such a drag.

Implementations

impl<'tcx, O> Obligation<'tcx, O>

pub fn new( tcx: TyCtxt<'tcx>, cause: ObligationCause<'tcx>, param_env: ParamEnv<'tcx>, predicate: impl ToPredicate<'tcx, O> ) -> Obligation<'tcx, O>

pub fn set_depth_from_parent(&mut self, parent_depth: usize)

We often create nested obligations without setting the correct depth.

To deal with this evaluate and fulfill explicitly update the depth of nested obligations using this function.

pub fn with_depth( tcx: TyCtxt<'tcx>, cause: ObligationCause<'tcx>, recursion_depth: usize, param_env: ParamEnv<'tcx>, predicate: impl ToPredicate<'tcx, O> ) -> Obligation<'tcx, O>

pub fn misc( tcx: TyCtxt<'tcx>, span: Span, body_id: LocalDefId, param_env: ParamEnv<'tcx>, trait_ref: impl ToPredicate<'tcx, O> ) -> Obligation<'tcx, O>

pub fn with<P>( &self, tcx: TyCtxt<'tcx>, value: impl ToPredicate<'tcx, P> ) -> Obligation<'tcx, P>