Module rustc_infer::infer::error_reporting::nice_region_error

Re-exports

pub use find_anon_type::find_anon_type;

Modules

different_lifetimes 🔒

Error Reporting for Anonymous Region Lifetime Errors where both the regions are anonymous.

find_anon_type

mismatched_static_lifetime 🔒

Error Reporting for when the lifetime for a type doesn’t match the impl selected for a predicate to hold.

named_anon_conflict 🔒

Error Reporting for Anonymous Region Lifetime Errors where one region is named and the other is anonymous.

placeholder_error 🔒

static_impl_trait 🔒

Error Reporting for static impl Traits.

trait_impl_difference 🔒

Error Reporting for impl items that do not match the obligations from their trait.

util 🔒

Helper functions corresponding to lifetime errors due to anonymous regions.

Structs

HirTraitObjectVisitor

Collect all hir::Ty<'_> Spans for trait objects with an implicit lifetime.

NiceRegionError

TraitObjectVisitor

Collect all the trait objects in a type that could have received an implicit 'static lifetime.

Functions

find_param_with_region

suggest_adding_lifetime_params

Currently only used in rustc_borrowck, probably should be removed in favour of public_errors::AddLifetimeParamsSuggestion

suggest_new_region_bound