Struct rustc_infer::infer::region_constraints::RegionConstraintCollector

pub struct RegionConstraintCollector<'a, 'tcx> {
    storage: &'a mut RegionConstraintStorage<'tcx>,
    undo_log: &'a mut InferCtxtUndoLogs<'tcx>,
}

Fields
storage: &'a mut RegionConstraintStorage<'tcx>undo_log: &'a mut InferCtxtUndoLogs<'tcx>
Implementations

impl<'tcx> RegionConstraintCollector<'_, 'tcx>

pub fn leak_check(
&mut self,
tcx: TyCtxt<'tcx>,
overly_polymorphic: bool,
max_universe: UniverseIndex,
snapshot: &CombinedSnapshot<'_, 'tcx>
) -> RelateResult<'tcx, ()>

Searches new universes created during snapshot, looking for placeholders that may “leak” out from the universes they are contained in. If any leaking placeholders are found, then an Err is returned (typically leading to the snapshot being reversed).

The leak check used to be the only way we had to handle higher-ranked obligations. Now that we have integrated universes into the region solvers, this is no longer the case, but we retain the leak check for backwards compatibility purposes. In particular, it lets us make “early” decisions about whether a region error will be reported that are used in coherence and elsewhere – see #56105 and #59490 for more details. The eventual fate of the leak checker is not yet settled.

The leak checker works by searching for the following error patterns:

P1: P2, where P1 != P2

P1: R, where R is in some universe that cannot name P1

The idea here is that each of these patterns represents something that the region solver would eventually report as an error, so we can detect the error early. There is a fly in the ointment, though, in that this is not entirely true. In particular, in the future, we may extend the environment with implied bounds or other info about how placeholders relate to regions in outer universes. In that case, P1: R for example might become solvable.

Summary of the implementation

The leak checks as follows. First, we construct a graph where R2: R1 implies R2 -> R1, and we compute the SCCs.

For each SCC S, we compute:

what placeholder P it must be equal to, if any

if there are multiple placeholders that must be equal, report an error because P1: P2

the minimum universe of its constituents

Then we walk the SCCs in dependency order and compute

what placeholder they must outlive transitively

if they must also be equal to a placeholder, report an error because P1: P2

minimum universe U of all SCCs they must outlive

if they must also be equal to a placeholder P, and U cannot name P, report an error, as that indicates P: R and R is in an incompatible universe

Historical note

Older variants of the leak check used to report errors for these patterns, but we no longer do:

R: P1, even if R cannot name P1, because R = ’static is a valid sol’n

R: P1, R: P2, as above

impl<'tcx> RegionConstraintCollector<'_, 'tcx>

pub fn num_region_vars(&self) -> usize

pub fn region_constraint_data(&self) -> &RegionConstraintData<'tcx>

pub fn into_infos_and_data(self) -> (VarInfos, RegionConstraintData<'tcx>)

Once all the constraints have been gathered, extract out the final data.

Not legal during a snapshot.

pub fn take_and_reset_data(&mut self) -> RegionConstraintData<'tcx>

Takes (and clears) the current set of constraints. Note that the set of variables remains intact, but all relationships between them are reset. This is used during NLL checking to grab the set of constraints that arose from a particular operation.

We don’t want to leak relationships between variables between points because just because (say) r1 == r2 was true at some point P in the graph doesn’t imply that it will be true at some other point Q, in NLL.

Not legal during a snapshot.

pub(super) fn data(&self) -> &RegionConstraintData<'tcx>

pub(super) fn start_snapshot(&mut self) -> RegionSnapshot

pub(super) fn rollback_to(&mut self, snapshot: RegionSnapshot)

pub(super) fn new_region_var(
&mut self,
universe: UniverseIndex,
origin: RegionVariableOrigin
) -> RegionVid

pub(super) fn var_universe(&self, vid: RegionVid) -> UniverseIndex

Returns the universe for the given variable.

pub(super) fn var_origin(&self, vid: RegionVid) -> RegionVariableOrigin

Returns the origin for the given variable.

fn add_constraint(
&mut self,
constraint: Constraint<'tcx>,
origin: SubregionOrigin<'tcx>
)

fn add_verify(&mut self, verify: Verify<'tcx>)

pub(super) fn add_given(&mut self, sub: Region<'tcx>, sup: RegionVid)

pub(super) fn make_eqregion(
&mut self,
origin: SubregionOrigin<'tcx>,
sub: Region<'tcx>,
sup: Region<'tcx>
)

pub(super) fn member_constraint(
&mut self,
key: OpaqueTypeKey<'tcx>,
definition_span: Span,
hidden_ty: Ty<'tcx>,
member_region: Region<'tcx>,
choice_regions: &Lrc<Vec<Region<'tcx>>>
)

pub(super) fn make_subregion(
&mut self,
origin: SubregionOrigin<'tcx>,
sub: Region<'tcx>,
sup: Region<'tcx>
)

pub(super) fn verify_generic_bound(
&mut self,
origin: SubregionOrigin<'tcx>,
kind: GenericKind<'tcx>,
sub: Region<'tcx>,
bound: VerifyBound<'tcx>
)

pub(super) fn lub_regions(
&mut self,
tcx: TyCtxt<'tcx>,
origin: SubregionOrigin<'tcx>,
a: Region<'tcx>,
b: Region<'tcx>
) -> Region<'tcx>

pub(super) fn glb_regions(
&mut self,
tcx: TyCtxt<'tcx>,
origin: SubregionOrigin<'tcx>,
a: Region<'tcx>,
b: Region<'tcx>
) -> Region<'tcx>

pub(super) fn opportunistic_resolve_var(&mut self, rid: RegionVid) -> RegionVid

Resolves the passed RegionVid to the root RegionVid in the unification table

pub fn opportunistic_resolve_region(
&mut self,
tcx: TyCtxt<'tcx>,
region: Region<'tcx>
) -> Region<'tcx>

If the Region is a ReVar, then resolves it either to the root value in the unification table, if it exists, or to the root ReVar in the table. If the Region is not a ReVar, just returns the Region itself.

fn combine_map(
&mut self,
t: CombineMapType
) -> &mut FxHashMap<TwoRegions<'tcx>, RegionVid>

fn combine_vars(
&mut self,
tcx: TyCtxt<'tcx>,
t: CombineMapType,
a: Region<'tcx>,
b: Region<'tcx>,
origin: SubregionOrigin<'tcx>
) -> Region<'tcx>

pub fn universe(&self, region: Region<'tcx>) -> UniverseIndex

pub fn vars_since_snapshot(
&self,
value_count: usize
) -> (Range<RegionVid>, Vec<RegionVariableOrigin>)

pub fn region_constraints_added_in_snapshot(
&self,
mark: &Snapshot<'tcx>
) -> Option<bool>

See InferCtxt::region_constraints_added_in_snapshot.

fn unification_table(
&mut self
) -> UnificationTable<InPlace<RegionVidKey<'tcx>, &'_ mut UnificationStorage<RegionVidKey<'tcx>>, &'_ mut InferCtxtUndoLogs<'tcx>>>