Struct rustc_infer::infer::InferCtxtBuilder

pub struct InferCtxtBuilder<'tcx> {
    tcx: TyCtxt<'tcx>,
    defining_use_anchor: DefiningAnchor,
    considering_regions: bool,
    skip_leak_check: bool,
    intercrate: bool,
    next_trait_solver: bool,
}

Used to configure inference contexts before their creation.

Fields
tcx: TyCtxt<'tcx>defining_use_anchor: DefiningAnchorconsidering_regions: boolskip_leak_check: boolintercrate: bool

Whether we are in coherence mode.

next_trait_solver: bool

Whether we should use the new trait solver in the local inference context, which affects things like which solver is used in predicate_may_hold.

Implementations

impl<'tcx> InferCtxtBuilder<'tcx>

pub fn with_opaque_type_inference( self, defining_use_anchor: DefiningAnchor ) -> Self

Whenever the InferCtxt should be able to handle defining uses of opaque types, you need to call this function. Otherwise the opaque type will be treated opaquely.

It is only meant to be called in two places, for typeck (via Inherited::build) and for the inference context used in mir borrowck.

pub fn with_next_trait_solver(self, next_trait_solver: bool) -> Self

pub fn intercrate(self, intercrate: bool) -> Self

pub fn ignoring_regions(self) -> Self

pub fn skip_leak_check(self, skip_leak_check: bool) -> Self

pub fn build_with_canonical<T>( &mut self, span: Span, canonical: &Canonical<'tcx, T> ) -> (InferCtxt<'tcx>, T, CanonicalVarValues<'tcx>)where T: TypeFoldable<TyCtxt<'tcx>>,

Given a canonical value C as a starting point, create an inference context that contains each of the bound values within instantiated as a fresh variable. The f closure is invoked with the new infcx, along with the instantiated value V and a substitution S. This substitution S maps from the bound values in C to their instantiated values in V (in other words, S(C) = V).

pub fn build(&mut self) -> InferCtxt<'tcx>