rustc_infer::traits
Enum ObligationCauseCode

pub enum ObligationCauseCode<'tcx> {

    Misc,
    SliceOrArrayElem,
    TupleElem,
    WhereClause(DefId, Span),
    WhereClauseInExpr(DefId, Span, HirId, usize),
    ReferenceOutlivesReferent(Ty<'tcx>),
    ObjectTypeBound(Ty<'tcx>, Region<'tcx>),
    Coercion {
        source: Ty<'tcx>,
        target: Ty<'tcx>,
    },
    AssignmentLhsSized,
    TupleInitializerSized,
    StructInitializerSized,
    VariableType(HirId),
    SizedArgumentType(Option<HirId>),
    SizedReturnType,
    SizedCallReturnType,
    SizedYieldType,
    InlineAsmSized,
    SizedClosureCapture(LocalDefId),
    SizedCoroutineInterior(LocalDefId),
    RepeatElementCopy {
        is_constable: IsConstable,
        elt_type: Ty<'tcx>,
        elt_span: Span,
        elt_stmt_span: Span,
    },
    FieldSized {
        adt_kind: AdtKind,
        span: Span,
        last: bool,
    },
    ConstSized,
    SharedStatic,
    BuiltinDerived(DerivedCause<'tcx>),
    ImplDerived(Box<ImplDerivedCause<'tcx>>),
    WellFormedDerived(DerivedCause<'tcx>),
    FunctionArg {
        arg_hir_id: HirId,
        call_hir_id: HirId,
        parent_code: InternedObligationCauseCode<'tcx>,
    },
    CompareImplItem {
        impl_item_def_id: LocalDefId,
        trait_item_def_id: DefId,
        kind: AssocKind,
    },
    CheckAssociatedTypeBounds {
        impl_item_def_id: LocalDefId,
        trait_item_def_id: DefId,
    },
    ExprAssignable,
    MatchExpressionArm(Box<MatchExpressionArmCause<'tcx>>),
    Pattern {
        span: Option<Span>,
        root_ty: Ty<'tcx>,
        origin_expr: bool,
    },
    IfExpression(Box<IfExpressionCause<'tcx>>),
    IfExpressionWithNoElse,
    MainFunctionType,
    StartFunctionType,
    LangFunctionType(Symbol),
    IntrinsicType,
    LetElse,
    MethodReceiver,
    UnifyReceiver(Box<UnifyReceiverContext<'tcx>>),
    ReturnNoExpression,
    ReturnValue(HirId),
    OpaqueReturnType(Option<(Ty<'tcx>, HirId)>),
    BlockTailExpression(HirId, MatchSource),
    TrivialBound,
    AwaitableExpr(HirId),
    ForLoopIterator,
    QuestionMark,
    WellFormed(Option<WellFormedLoc>),
    MatchImpl(ObligationCause<'tcx>, DefId),
    BinOp {
        lhs_hir_id: HirId,
        rhs_hir_id: Option<HirId>,
        rhs_span: Option<Span>,
        rhs_is_lit: bool,
        output_ty: Option<Ty<'tcx>>,
    },
    AscribeUserTypeProvePredicate(Span),
    RustCall,
    DropImpl,
    ConstParam(Ty<'tcx>),
    TypeAlias(InternedObligationCauseCode<'tcx>, Span, DefId),

}

Variants

Misc

Not well classified or should be obvious from the span.

SliceOrArrayElem

A slice or array is WF only if T: Sized.

TupleElem

A tuple is WF only if its middle elements are Sized.

WhereClause(DefId, Span)

Represents a clause that comes from a specific item. The span corresponds to the clause.

WhereClauseInExpr(DefId, Span, HirId, usize)

Like WhereClause, but also identifies the expression which requires the where clause to be proven, and also identifies the index of the predicate in the predicates_of list of the item.

ReferenceOutlivesReferent(Ty<'tcx>)

A type like &'a T is WF only if T: 'a.

ObjectTypeBound(Ty<'tcx>, Region<'tcx>)

A type like Box<Foo<'a> + 'b> is WF only if 'b: 'a.

Coercion

Obligation incurred due to a coercion.

Fields

source: Ty<'tcx>

target: Ty<'tcx>

AssignmentLhsSized

Various cases where expressions must be Sized / Copy / etc. L = X implies that L is Sized.

TupleInitializerSized

(x1, .., xn) must be Sized.

StructInitializerSized

S { ... } must be Sized.

VariableType(HirId)

Type of each variable must be Sized.

SizedArgumentType(Option<HirId>)

Argument type must be Sized.

SizedReturnType

Return type must be Sized.

SizedCallReturnType

Return type of a call expression must be Sized.

SizedYieldType

Yield type must be Sized.

InlineAsmSized

Inline asm operand type must be Sized.

SizedClosureCapture(LocalDefId)

Captured closure type must be Sized.

SizedCoroutineInterior(LocalDefId)

Types live across coroutine yields must be Sized.

RepeatElementCopy

[expr; N] requires type_of(expr): Copy.

Fields

is_constable: IsConstable

If element is a const fn or const ctor we display a help message suggesting to move it to a new const item while saying that T doesn’t implement Copy.

elt_type: Ty<'tcx>

elt_span: Span

elt_stmt_span: Span

Span of the statement/item in which the repeat expression occurs. We can use this to place a const declaration before it

FieldSized

Types of fields (other than the last, except for packed structs) in a struct must be sized.

Fields

adt_kind: AdtKind

span: Span

last: bool

ConstSized

Constant expressions must be sized.

SharedStatic

static items must have Sync type.

BuiltinDerived(DerivedCause<'tcx>)

Derived obligation (i.e. theoretical where clause) on a built-in implementation like Copy or Sized.

ImplDerived(Box<ImplDerivedCause<'tcx>>)

Derived obligation (i.e. where clause) on an user-provided impl or a trait alias.

WellFormedDerived(DerivedCause<'tcx>)

Derived obligation for WF goals.

FunctionArg

Derived obligation refined to point at a specific argument in a call or method expression.

Fields

arg_hir_id: HirId

The node of the relevant argument in the function call.

call_hir_id: HirId

The node of the function call.

parent_code: InternedObligationCauseCode<'tcx>

The obligation introduced by this argument.

CompareImplItem

Error derived when checking an impl item is compatible with its corresponding trait item’s definition

Fields

impl_item_def_id: LocalDefId

trait_item_def_id: DefId

kind: AssocKind

CheckAssociatedTypeBounds

Checking that the bounds of a trait’s associated type hold for a given impl

Fields

impl_item_def_id: LocalDefId

trait_item_def_id: DefId

ExprAssignable

Checking that this expression can be assigned to its target.

MatchExpressionArm(Box<MatchExpressionArmCause<'tcx>>)

Computing common supertype in the arms of a match expression

Pattern

Type error arising from type checking a pattern against an expected type.

Fields

span: Option<Span>

The span of the scrutinee or type expression which caused the root_ty type.

root_ty: Ty<'tcx>

The root expected type induced by a scrutinee or type expression.

origin_expr: bool

Whether the Span came from an expression or a type expression.

IfExpression(Box<IfExpressionCause<'tcx>>)

Computing common supertype in an if expression

IfExpressionWithNoElse

Computing common supertype of an if expression with no else counter-part

MainFunctionType

main has wrong type

StartFunctionType

start has wrong type

LangFunctionType(Symbol)

language function has wrong type

IntrinsicType

Intrinsic has wrong type

LetElse

A let else block does not diverge

MethodReceiver

Method receiver

UnifyReceiver(Box<UnifyReceiverContext<'tcx>>)

ReturnNoExpression

return with no expression

ReturnValue(HirId)

return with an expression

OpaqueReturnType(Option<(Ty<'tcx>, HirId)>)

Opaque return type of this function

BlockTailExpression(HirId, MatchSource)

Block implicit return

TrivialBound

#[feature(trivial_bounds)] is not enabled

AwaitableExpr(HirId)

ForLoopIterator

QuestionMark

WellFormed(Option<WellFormedLoc>)

Well-formed checking. If a WellFormedLoc is provided, then it will be used to perform HIR-based wf checking after an error occurs, in order to generate a more precise error span. This is purely for diagnostic purposes - it is always correct to use Misc instead, or to specify WellFormed(None).

MatchImpl(ObligationCause<'tcx>, DefId)

From match_impl. The cause for us having to match an impl, and the DefId we are matching against.

BinOp

Fields

lhs_hir_id: HirId

rhs_hir_id: Option<HirId>

rhs_span: Option<Span>

rhs_is_lit: bool

output_ty: Option<Ty<'tcx>>

AscribeUserTypeProvePredicate(Span)

RustCall

DropImpl

Obligations to prove that a std::ops::Drop impl is not stronger than the ADT it’s being implemented for.

ConstParam(Ty<'tcx>)

Requirement for a const N: Ty to implement Ty: ConstParamTy

TypeAlias(InternedObligationCauseCode<'tcx>, Span, DefId)

Obligations emitted during the normalization of a weak type alias.

Implementations

impl<'tcx> ObligationCauseCode<'tcx>

pub fn peel_derives(&self) -> &ObligationCauseCode<'tcx>

Returns the base obligation, ignoring derived obligations.

pub fn peel_derives_with_predicate( &self, ) -> (&ObligationCauseCode<'tcx>, Option<Binder<TyCtxt<'tcx>, TraitPredicate<TyCtxt<'tcx>>>>)

Returns the base obligation and the base trait predicate, if any, ignoring derived obligations.

pub fn parent( &self, ) -> Option<(&ObligationCauseCode<'tcx>, Option<Binder<TyCtxt<'tcx>, TraitPredicate<TyCtxt<'tcx>>>>)>

pub fn peel_match_impls(&self) -> &ObligationCauseCode<'tcx>