Type Definition rustc_infer::traits::TraitObligation

pub type TraitObligation<'tcx> = Obligation<'tcx, PolyTraitPredicate<'tcx>>;

Implementations

impl<'tcx> TraitObligation<'tcx>

pub fn is_const(&self) -> bool

Returns true if the trait predicate is considered const in its ParamEnv.

pub fn derived_cause(
&self,
variant: impl FnOnce(DerivedObligationCause<'tcx>) -> ObligationCauseCode<'tcx>
) -> ObligationCause<'tcx>

impl<'tcx> TraitObligation<'tcx>

pub fn polarity(&self) -> ImplPolarity

pub fn self_ty(&self) -> Binder<'tcx, Ty<'tcx>>

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference’s “Type Layout” chapter for details on type layout guarantees.

Size: 72 bytes